# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DeciFlow API",
        "description": "API documentation for DeciFlow - Purchase Request Approval System",
        "contact": {
            "email": "support@deciflow.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/requests/{id}/attachments": {
            "get": {
                "tags": [
                    "Attachments"
                ],
                "summary": "List attachments",
                "description": "Get all attachments for a request",
                "operationId": "2d8f08aad73f4a6dd5efaac42b652a6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of attachments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Request not found"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Attachments"
                ],
                "summary": "Upload attachment",
                "description": "Upload a file attachment to a request",
                "operationId": "5b5dde6ce716551c4202ff58e15a9e09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "File to upload (max 10MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Attachment uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/attachments/{id}": {
            "get": {
                "tags": [
                    "Attachments"
                ],
                "summary": "Download attachment",
                "description": "Download a file attachment",
                "operationId": "d231512cd65e91bbb246e817c47170e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Attachment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File download",
                        "content": {
                            "application/octet-stream": {}
                        }
                    },
                    "404": {
                        "description": "Attachment not found"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Attachments"
                ],
                "summary": "Delete attachment",
                "description": "Delete a file attachment",
                "operationId": "9d090bf09bcab12eef937baddc7176c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Attachment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attachment deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Attachment deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Attachment not found"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/audit": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "List all audit logs",
                "description": "Get all audit logs across all requests (super admin only)",
                "operationId": "e93a854a4cee132b49aa967ce67ac4ff",
                "responses": {
                    "200": {
                        "description": "Paginated list of audit logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - super admin only"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/requests/{id}/audit": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "Get request audit trail",
                "description": "Get all audit logs for a specific request",
                "operationId": "6460d639ed89ceef8f6f2b805c004668",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of audit logs for the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}